Add basic password policy validation in Installer
This is a little bit of a hack, ideally this should be in the backend and the frontend should consume the error message.
But the setup is a separare step so the error message is shown too late.

As a workaround for the meanwhile, enforce some password in the UI.